+++ src/Planners/src/UnicycleTrajectoryGenerator.cpp
@@ -99,6 +99,10 @@ public:
 
     BipedalLocomotion::Planners::UnicycleTrajectoryPlanner unicycleTrajectoryPlanner;
 
+    manif::SE3d measuredLeftFootTransform;
+    manif::SE3d measuredRightFootTransform;
+    bool useMeasuredFeetTransforms;
+
     /**
      * ask for a new trajectory to the unicycle trajectory planner
      */
@@ -558,12 +562,24 @@ bool Planners::UnicycleTrajectoryGenerator::advance()
 
             std::chrono::nanoseconds initTimeTrajectory
                 = m_pImpl->time + m_pImpl->newTrajectoryMergeCounter * m_pImpl->parameters.dt;
+            manif::SE3d measuredTransform;
 
-            manif::SE3d measuredTransform = m_pImpl->trajectory.isLeftFootLastSwinging.front()
-                                                ? m_pImpl->trajectory.rightFootTransform.at(
-                                                    m_pImpl->newTrajectoryMergeCounter)
-                                                : m_pImpl->trajectory.leftFootTransform.at(
-                                                    m_pImpl->newTrajectoryMergeCounter);
+            if (m_pImpl->useMeasuredFeetTransforms)
+            {
+                measuredTransform = m_pImpl->trajectory.isLeftFootLastSwinging.front()
+                                        ? m_pImpl->measuredRightFootTransform
+                                        : m_pImpl->measuredLeftFootTransform;
+                m_pImpl->useMeasuredFeetTransforms = false; // once used, it is no more usable. User
+                                                            // should reset it again.
+            } else
+            {
+
+                measuredTransform = m_pImpl->trajectory.isLeftFootLastSwinging.front()
+                                        ? m_pImpl->trajectory.rightFootTransform.at(
+                                              m_pImpl->newTrajectoryMergeCounter)
+                                        : m_pImpl->trajectory.leftFootTransform.at(
+                                              m_pImpl->newTrajectoryMergeCounter);
+            }
 
             // ask for a new trajectory (and spawn an asynchronous thread to compute it)
             {
@@ -957,4 +973,12 @@ std::chrono::nanoseconds
 BipedalLocomotion::Planners::UnicycleTrajectoryGenerator::getSamplingTime() const
 {
     return m_pImpl->parameters.dt;
+};
+
+void BipedalLocomotion::Planners::UnicycleTrajectoryGenerator::setFeetTransform(
+    const manif::SE3d& leftFootTransform, const manif::SE3d& rightFootTransform)
+{
+    m_pImpl->measuredLeftFootTransform = leftFootTransform;
+    m_pImpl->measuredRightFootTransform = rightFootTransform;
+    m_pImpl->useMeasuredFeetTransforms = true;
 };